iter35 cluster-037 fix r1: 补 architect+quality demand
⟦AI:AUTO-LOOP⟧
