+++ src/Aevatar.Mainnet.Host.Api/Responses/ResponsesApiModels.cs
@@ -1,8 +1,6 @@
-using System.Diagnostics.CodeAnalysis;
-using System.Security.Cryptography;
-using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using Aevatar.GAgentService.Application.Responses;
 
 namespace Aevatar.Mainnet.Host.Api.Responses;
 
@@ -28,329 +26,6 @@ internal sealed record ResponsesCreateRequest
 
     [JsonPropertyName("tools")]
     public JsonElement Tools { get; init; }
-}
-
-internal sealed record NormalizedResponsesRequest(
-    string ResponseId,
-    string MessageItemId,
-    string Model,
-    string Prompt,
-    bool Stream,
-    string? PreviousResponseId,
-    double? Temperature,
-    int? MaxOutputTokens,
-    IReadOnlyList<ResponsesToolDeclaration> DeclaredTools,
-    IReadOnlyList<ResponsesToolResultInput> ToolResults);
-
-internal sealed record ResponsesToolDeclaration(
-    string Name,
-    string Description,
-    string ParametersJson,
-    string SchemaHash);
-
-internal sealed record ResponsesToolResultInput(
-    string CallId,
-    string Output,
-    string? SchemaHash);
-
-internal readonly record struct ResponsesRequestNormalizationResult(
-    NormalizedResponsesRequest? Request,
-    string? ErrorCode,
-    string? ErrorMessage)
-{
-    public bool Succeeded => Request != null && ErrorCode == null;
-
-    public static ResponsesRequestNormalizationResult Success(NormalizedResponsesRequest request) =>
-        new(request, null, null);
-
-    public static ResponsesRequestNormalizationResult Failed(string code, string message) =>
-        new(null, code, message);
-}
-
-internal static class ResponsesRequestNormalizer
-{
-    public static ResponsesRequestNormalizationResult Normalize(ResponsesCreateRequest request)
-    {
-        ArgumentNullException.ThrowIfNull(request);
-
-        var model = request.Model?.Trim();
-        if (string.IsNullOrWhiteSpace(model))
-            return ResponsesRequestNormalizationResult.Failed("model_required", "model is required.");
-
-        if (!TryExtractDeclaredTools(request.Tools, out var declaredTools, out var toolsError))
-            return ResponsesRequestNormalizationResult.Failed("invalid_tools", toolsError);
-
-        if (!TryExtractInput(request.Input, out var prompt, out var toolResults, out var inputError))
-            return ResponsesRequestNormalizationResult.Failed("invalid_input", inputError);
-
-        if (request.MaxOutputTokens is <= 0)
-        {
-            return ResponsesRequestNormalizationResult.Failed(
-                "invalid_max_output_tokens",
-                "max_output_tokens must be greater than zero when provided.");
-        }
-
-        var previousResponseId = NormalizeOptional(request.PreviousResponseId);
-
-        // OpenAI Responses spec pairs `function_call_output` items in `input` with
-        // `previous_response_id` so the server can match them to a pending tool call
-        // (#629 §13 continuation contract). But Anthropic→OpenAI translators (CC Switch,
-        // Codex when wrapping Claude Code) often forward Claude Code's prior tool-result
-        // turns in the `input` array WITHOUT propagating previous_response_id — they
-        // don't model OpenAI's server-side session. Treating that strictly returns
-        // `function_call_output requires previous_response_id` and the agent can't
-        // ever continue a multi-turn tool conversation.
-        //
-        // Resolution: when previous_response_id is absent, fold any function_call_output
-        // entries into the user prompt as historical context (with a synthetic
-        // `[tool_result …]` marker) and clear ToolResults. The continuation contract
-        // only kicks in when previous_response_id IS provided — that path is unchanged.
-        if (previousResponseId is null && toolResults.Count > 0)
-        {
-            var foldedSections = new List<string>();
-            if (!string.IsNullOrWhiteSpace(prompt))
-                foldedSections.Add(prompt);
-            foreach (var tr in toolResults)
-            {
-                var marker = $"[tool_result call_id={tr.CallId}]";
-                foldedSections.Add(string.IsNullOrWhiteSpace(tr.Output) ? marker : $"{marker} {tr.Output}");
-            }
-            prompt = string.Join("\n", foldedSections);
-            toolResults = [];
-        }
-
-        return ResponsesRequestNormalizationResult.Success(new NormalizedResponsesRequest(
-            ResponseId: ResponsesIds.NewResponseId(),
-            MessageItemId: ResponsesIds.NewMessageId(),
-            Model: model,
-            Prompt: prompt,
-            Stream: request.Stream == true,
-            PreviousResponseId: previousResponseId,
-            Temperature: request.Temperature,
-            MaxOutputTokens: request.MaxOutputTokens,
-            DeclaredTools: declaredTools,
-            ToolResults: toolResults));
-    }
-
-    private static string? NormalizeOptional(string? value)
-    {
-        var normalized = value?.Trim();
-        return string.IsNullOrWhiteSpace(normalized) ? null : normalized;
-    }
-
-    private static bool TryExtractInput(
-        JsonElement input,
-        [NotNullWhen(true)] out string? prompt,
-        out IReadOnlyList<ResponsesToolResultInput> toolResults,
-        [NotNullWhen(false)] out string? error)
-    {
-        var parts = new List<string>();
-        var results = new List<ResponsesToolResultInput>();
-        ExtractInput(input, parts, results);
-
-        prompt = string.Join("\n", parts.Select(static x => x.Trim()).Where(static x => x.Length > 0));
-        toolResults = results;
-        if (prompt.Length > 0 || results.Count > 0)
-        {
-            error = null;
-            return true;
-        }
-
-        error = "input must contain at least one text value.";
-        prompt = null;
-        toolResults = [];
-        return false;
-    }
-
-    private static void ExtractInput(
-        JsonElement element,
-        ICollection<string> parts,
-        ICollection<ResponsesToolResultInput> toolResults)
-    {
-        switch (element.ValueKind)
-        {
-            case JsonValueKind.String:
-                AddText(element.GetString(), parts);
-                return;
-
-            case JsonValueKind.Array:
-                foreach (var item in element.EnumerateArray())
-                    ExtractInput(item, parts, toolResults);
-                return;
-
-            case JsonValueKind.Object:
-                ExtractObjectInput(element, parts, toolResults);
-                return;
-        }
-    }
-
-    private static void ExtractObjectInput(
-        JsonElement element,
-        ICollection<string> parts,
-        ICollection<ResponsesToolResultInput> toolResults)
-    {
-        if (TryExtractToolResult(element, out var toolResult))
-        {
-            toolResults.Add(toolResult);
-            return;
-        }
-
-        if (element.TryGetProperty("text", out var text))
-        {
-            ExtractInput(text, parts, toolResults);
-            return;
-        }
-
-        if (element.TryGetProperty("content", out var content))
-        {
-            ExtractInput(content, parts, toolResults);
-            return;
-        }
-
-        if (element.TryGetProperty("input_text", out var inputText))
-        {
-            ExtractInput(inputText, parts, toolResults);
-        }
-    }
-
-    private static void AddText(string? value, ICollection<string> parts)
-    {
-        if (!string.IsNullOrWhiteSpace(value))
-            parts.Add(value);
-    }
-
-    private static bool TryExtractToolResult(
-        JsonElement element,
-        [NotNullWhen(true)] out ResponsesToolResultInput? toolResult)
-    {
-        toolResult = null;
-        var type = GetStringProperty(element, "type");
-        if (!string.Equals(type, "function_call_output", StringComparison.OrdinalIgnoreCase) &&
-            !string.Equals(type, "tool_result", StringComparison.OrdinalIgnoreCase))
-        {
-            return false;
-        }
-
-        var callId = GetStringProperty(element, "call_id")
-                     ?? GetStringProperty(element, "tool_call_id")
-                     ?? GetStringProperty(element, "id");
-        if (string.IsNullOrWhiteSpace(callId))
-            return false;
-
-        string? output = null;
-        if (element.TryGetProperty("output", out var outputElement))
-            output = ElementToPayloadString(outputElement);
-        else if (element.TryGetProperty("result", out var resultElement))
-            output = ElementToPayloadString(resultElement);
-
-        var schemaHash = GetStringProperty(element, "schema_hash")
-                         ?? GetStringProperty(element, "schemaHash");
-        toolResult = new ResponsesToolResultInput(
-            callId.Trim(),
-            output ?? string.Empty,
-            NormalizeOptional(schemaHash));
-        return true;
-    }
-
-    private static bool TryExtractDeclaredTools(
-        JsonElement tools,
-        out IReadOnlyList<ResponsesToolDeclaration> declaredTools,
-        [NotNullWhen(false)] out string? error)
-    {
-        declaredTools = [];
-        error = null;
-        if (tools.ValueKind is JsonValueKind.Undefined or JsonValueKind.Null)
-            return true;
-        if (tools.ValueKind != JsonValueKind.Array)
-        {
-            error = "tools must be an array when provided.";
-            return false;
-        }
-
-        var result = new List<ResponsesToolDeclaration>();
-        var toolIndex = -1;
-        foreach (var tool in tools.EnumerateArray())
-        {
-            toolIndex++;
-            if (tool.ValueKind != JsonValueKind.Object)
-            {
-                error = $"tool at index {toolIndex} must be an object.";
-                return false;
-            }
-
-            // OpenAI Responses API allows built-in tool declarations like
-            // `{type: "web_search_preview"}` / `{type: "file_search", ...}` /
-            // `{type: "code_interpreter", ...}` / `{type: "computer_use_preview", ...}`
-            // that don't carry a `function` block or `name`. They're routing hints to
-            // the model provider, not custom function definitions. aevatar's classifier
-            // only owns function-typed tools (forward / substitute / additive); silently
-            // pass over the rest so an OpenAI-compatible client (CC Switch, Codex,
-            // Cursor) can advertise built-ins without breaking the request — even
-            // though aevatar won't map them to a local handler and the model provider
-            // gets to decide what to do with them.
-            // OpenAI Responses API allows built-in tool declarations like
-            // `{type: "web_search_preview"}` / `{type: "file_search", ...}` /
-            // `{type: "code_interpreter", ...}` / `{type: "computer_use_preview", ...}`
-            // that don't carry a `function` block or `name`. They're routing hints to
-            // the model provider, not custom function definitions. aevatar's classifier
-            // only owns function-typed tools (forward / substitute / additive); silently
-            // pass over the rest so an OpenAI-compatible client (CC Switch, Codex,
-            // Cursor) can advertise built-ins without breaking the request — even
-            // though aevatar won't map them to a local handler and the model provider
-            // gets to decide what to do with them.
-            var toolType = GetStringProperty(tool, "type");
-            var isFunctionType = string.IsNullOrWhiteSpace(toolType) ||
-                                 string.Equals(toolType, "function", StringComparison.OrdinalIgnoreCase);
-            if (!isFunctionType)
-                continue;
-
-            var function = tool.TryGetProperty("function", out var functionElement) &&
-                           functionElement.ValueKind == JsonValueKind.Object
-                ? functionElement
-                : tool;
-            var name = GetStringProperty(function, "name");
-            if (string.IsNullOrWhiteSpace(name))
-            {
-                error = $"function tool at index {toolIndex} requires a non-empty name.";
-                return false;
-            }
-
-            var description = GetStringProperty(function, "description") ?? string.Empty;
-            var parametersJson = function.TryGetProperty("parameters", out var parameters)
-                ? ElementToPayloadString(parameters)
-                : """{"type":"object","properties":{}}""";
-            result.Add(new ResponsesToolDeclaration(
-                name.Trim(),
-                description,
-                parametersJson,
-                ResponsesToolSchemaHashes.Compute(parametersJson)));
-        }
-
-        declaredTools = result;
-        return true;
-    }
-
-    private static string? GetStringProperty(JsonElement element, string propertyName)
-    {
-        if (!element.TryGetProperty(propertyName, out var property))
-            return null;
-        return property.ValueKind == JsonValueKind.String ? property.GetString() : null;
-    }
-
-    private static string ElementToPayloadString(JsonElement element) =>
-        element.ValueKind == JsonValueKind.String
-            ? element.GetString() ?? string.Empty
-            : element.GetRawText();
-
-}
-
-internal static class ResponsesToolSchemaHashes
-{
-    public static string Compute(string parametersJson)
-    {
-        var bytes = SHA256.HashData(Encoding.UTF8.GetBytes(parametersJson));
-        return Convert.ToHexString(bytes).ToLowerInvariant();
-    }
 }
 
 internal sealed record ResponsesApiErrorResponse
@@ -648,61 +323,4 @@ internal sealed record ResponsesModelEntry
     [JsonPropertyName("description")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public string? Description { get; init; }
-}
-
-/// <summary>Splits an OpenRouter-style `vendor/model` identifier. Vendor is preserved only when it
-/// looks like a NyxID service slug (lowercase + digits + hyphens, length 2-64); otherwise the whole
-/// string is treated as a bare model name (e.g. for Anthropic-style namespacing the LLM provider
-/// may want intact).
-/// Gateway-routed models are emitted bare by the catalog, so a prefix is only ever produced for
-/// UserService / ProxyService routes that resolve to `/api/v1/proxy/s/{slug}`. A client that
-/// invents an unknown slug gets a clean NyxID 404 — fail-closed; we don't validate against the
-/// catalog here to keep `/v1/responses` off the catalog HTTP critical path.</summary>
-internal static class ResponsesModelRouteParser
-{
-    public static ResponsesModelRoute Parse(string model)
-    {
-        ArgumentException.ThrowIfNullOrWhiteSpace(model);
-        var trimmed = model.Trim();
-        var slashIndex = trimmed.IndexOf('/');
-        if (slashIndex <= 0 || slashIndex >= trimmed.Length - 1)
-            return new ResponsesModelRoute(null, trimmed);
-
-        var prefix = trimmed[..slashIndex];
-        var rest = trimmed[(slashIndex + 1)..];
-        return LooksLikeSlug(prefix)
-            ? new ResponsesModelRoute(prefix, rest)
-            : new ResponsesModelRoute(null, trimmed);
-    }
-
-    private static bool LooksLikeSlug(string value)
-    {
-        if (value.Length is < 2 or > 64) return false;
-        if (!char.IsAsciiLetterLower(value[0])) return false;
-        foreach (var c in value)
-        {
-            if (!(char.IsAsciiLetterLower(c) || char.IsAsciiDigit(c) || c == '-'))
-                return false;
-        }
-        return true;
-    }
-}
-
-internal readonly record struct ResponsesModelRoute(string? RouteSlug, string Model);
-
-internal static class ResponsesIds
-{
-    public static string NewResponseId() => "resp_" + NewOpaqueId();
-
-    public static string NewMessageId() => "msg_" + NewOpaqueId();
-
-    public static string NewOpaqueId()
-    {
-        Span<byte> bytes = stackalloc byte[16];
-        RandomNumberGenerator.Fill(bytes);
-        return Convert.ToBase64String(bytes)
-            .TrimEnd('=')
-            .Replace('+', '-')
-            .Replace('/', '_');
-    }
 }
