+++ src/Aevatar.Mainnet.Host.Api/Responses/ResponsesEndpoints.cs
@@ -1,25 +1,20 @@
-using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using System.Text.Json;
 using Aevatar.AI.Abstractions.LLMProviders;
 using Aevatar.ChatRouting.Abstractions;
-using Aevatar.ChatRouting.Core;
 using Aevatar.Foundation.Abstractions.Connectors;
 using Aevatar.GAgentService.Abstractions;
 using Aevatar.GAgentService.Abstractions.Ports;
-using Aevatar.GAgentService.Abstractions.Queries;
 using Aevatar.GAgentService.Abstractions.Responses;
 using Aevatar.GAgentService.Abstractions.ScopeGAgents;
 using Aevatar.GAgentService.Abstractions.Services;
 using Aevatar.GAgentService.Application.Responses;
 using Aevatar.GAgents.Channel.Runtime;
 using Aevatar.Presentation.AGUI;
-using Google.Protobuf.WellKnownTypes;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Routing;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
 namespace Aevatar.Mainnet.Host.Api.Responses;
@@ -49,10 +44,20 @@ internal static partial class ResponsesApiEndpoints
     internal static async Task<IResult> HandleCreateResponseAsync(
         HttpContext http,
         ResponsesCreateRequest request,
+        [FromServices] IResponsesCommandFacade commandFacade,
+        [FromServices] ITeamEntryMemberResolver teamEntryMemberResolver,
+        [FromServices] IMemberPublishedServiceResolver memberPublishedServiceResolver,
+        [FromServices] IStaticGAgentStreamInvocationPort<AGUIEvent> staticGAgentStreamInvocationPort,
+        [FromServices] ILoggerFactory loggerFactory,
         CancellationToken ct)
     {
         ArgumentNullException.ThrowIfNull(http);
         ArgumentNullException.ThrowIfNull(request);
+        ArgumentNullException.ThrowIfNull(commandFacade);
+        ArgumentNullException.ThrowIfNull(teamEntryMemberResolver);
+        ArgumentNullException.ThrowIfNull(memberPublishedServiceResolver);
+        ArgumentNullException.ThrowIfNull(staticGAgentStreamInvocationPort);
+        ArgumentNullException.ThrowIfNull(loggerFactory);
 
         var bearerToken = ExtractBearerToken(http);
         if (string.IsNullOrWhiteSpace(bearerToken))
@@ -61,32 +66,71 @@ internal static partial class ResponsesApiEndpoints
                 "authentication_required",
                 "Authorization bearer token is required.");
 
-        var facade = ActivatorUtilities.CreateInstance<ResponsesCommandFacade>(http.RequestServices);
-        return await facade.CreateAsync(http, request, bearerToken, ct);
-    }
+        var result = await commandFacade.CreateAsync(ToCommandRequest(request), bearerToken, ct);
+        if (result.Error is not null)
+            return ToErrorResult(result.Error.StatusCode, result.Error.Code, result.Error.Message);
 
-    private static string LogAndCorrelate(
-        ILogger logger,
-        Exception ex,
-        string stage,
-        string responseId)
-    {
-        var correlation = Guid.NewGuid().ToString("N")[..16];
-        logger.LogError(
-            ex,
-            "Responses {Stage} failure for {ResponseId} (correlation {Correlation}).",
-            stage,
-            responseId,
-            correlation);
-        return correlation;
+        var logger = loggerFactory.CreateLogger("Aevatar.Mainnet.Host.Api.Responses");
+        if (result.Forward?.Action.ForwardToTeam is not null)
+        {
+            return await HandleForwardToTeamAsync(
+                http,
+                result.Forward.Normalized,
+                result.Forward.CallerScope,
+                result.Forward.Action.ForwardToTeam,
+                teamEntryMemberResolver,
+                staticGAgentStreamInvocationPort,
+                logger,
+                ct);
+        }
+
+        if (result.Forward?.Action.ForwardToGagent is not null)
+        {
+            return await HandleForwardToGAgentAsync(
+                http,
+                result.Forward.Normalized,
+                result.Forward.CallerScope,
+                result.Forward.Action.ForwardToGagent,
+                memberPublishedServiceResolver,
+                staticGAgentStreamInvocationPort,
+                logger,
+                ct);
+        }
+
+        if (result.StreamPlan is not null)
+        {
+            await WriteStreamResponseAsync(
+                http.Response,
+                commandFacade,
+                result.StreamPlan,
+                ct);
+            return Results.Empty;
+        }
+
+        if (result.Completed is not null)
+        {
+            return Results.Json(
+                BuildCompletedResponse(
+                    result.Completed.Normalized,
+                    result.Completed.CreatedAt,
+                    result.Completed.CompletedAt,
+                    result.Completed.OutputText,
+                    result.Completed.ForwardedToolCalls,
+                    result.Completed.Usage is null ? null : MapUsage(result.Completed.Usage)),
+                statusCode: StatusCodes.Status200OK);
+        }
+
+        throw new InvalidOperationException("Responses command facade returned no result.");
     }
 
     internal static async Task<IResult> HandleCancelResponseAsync(
         HttpContext http,
         [FromRoute] string id,
+        [FromServices] IResponsesCommandFacade commandFacade,
         CancellationToken ct)
     {
         ArgumentNullException.ThrowIfNull(http);
+        ArgumentNullException.ThrowIfNull(commandFacade);
 
         var responseId = id?.Trim();
         if (string.IsNullOrWhiteSpace(responseId))
@@ -104,26 +148,37 @@ internal static partial class ResponsesApiEndpoints
                 "authentication_required",
                 "Authorization bearer token is required.");
 
-        var facade = ActivatorUtilities.CreateInstance<ResponsesCommandFacade>(http.RequestServices);
-        return await facade.CancelAsync(responseId, bearerToken, http, ct);
+        var result = await commandFacade.CancelAsync(responseId, bearerToken, ct);
+        if (result.Error is not null)
+            return ToErrorResult(result.Error.StatusCode, result.Error.Code, result.Error.Message);
+
+        return Results.Json(new
+        {
+            id = result.ResponseId,
+            @object = "response",
+            status = "cancelled",
+            cancelled_at = result.CancelledAt,
+        }, JsonOptions, statusCode: StatusCodes.Status200OK);
     }
+
+    private static ResponsesCommandRequest ToCommandRequest(ResponsesCreateRequest request) =>
+        new(
+            request.Model,
+            request.Input,
+            request.Stream,
+            request.PreviousResponseId,
+            request.Temperature,
+            request.MaxOutputTokens,
+            request.Tools);
 
     private static async Task WriteStreamResponseAsync(
         HttpResponse response,
-        ILLMProviderFactory providerFactory,
-        IResponsesCompletionApplicationService completionService,
-        ILlmSessionRegistrationPort responseSessionRegistrationPort,
-        ILogger logger,
-        LlmSessionRegistrationResult responseSession,
-        LLMRequest request,
-        IReadOnlyDictionary<string, string> toolContextMetadata,
-        NormalizedResponsesRequest normalized,
-        LlmSessionSnapshot? previousSnapshot,
-        ResponsesToolClassification toolClassification,
-        DateTimeOffset createdAtOffset,
+        IResponsesCommandFacade commandFacade,
+        ResponsesCreateCommandPlan plan,
         CancellationToken ct)
     {
-        var createdAt = createdAtOffset.ToUnixTimeSeconds();
+        var normalized = plan.Normalized;
+        var createdAt = plan.CreatedAt.ToUnixTimeSeconds();
         response.StatusCode = StatusCodes.Status200OK;
         response.ContentType = "text/event-stream; charset=utf-8";
         response.Headers.CacheControl = "no-store";
@@ -132,222 +187,139 @@ internal static partial class ResponsesApiEndpoints
         await response.StartAsync(ct);
 
         var sequenceNumber = 0;
-        TokenUsage? usage = null;
+        var createdResponse = BuildCreatedResponse(normalized, createdAt);
+        await WriteSseFrameAsync(
+            response,
+            "response.created",
+            new
+            {
+                type = "response.created",
+                response = createdResponse,
+                sequence_number = ++sequenceNumber,
+            },
+            ct);
 
-        try
+        var outputItem = BuildOutputMessage(normalized.MessageItemId, "in_progress", text: null);
+        await WriteSseFrameAsync(
+            response,
+            "response.output_item.added",
+            new
+            {
+                type = "response.output_item.added",
+                output_index = 0,
+                item = outputItem,
+                sequence_number = ++sequenceNumber,
+            },
+            ct);
+
+        var completion = await commandFacade.StreamAsync(
+            plan,
+            async (delta, token) =>
+            {
+                await WriteSseFrameAsync(
+                    response,
+                    "response.output_text.delta",
+                    new
+                    {
+                        type = "response.output_text.delta",
+                        item_id = normalized.MessageItemId,
+                        output_index = 0,
+                        content_index = 0,
+                        delta,
+                        sequence_number = ++sequenceNumber,
+                    },
+                    token);
+            },
+            ct);
+
+        if (completion.Error is not null)
         {
-            var provider = providerFactory.GetDefault();
-            var createdResponse = BuildCreatedResponse(normalized, createdAt);
-            await WriteSseFrameAsync(
+            await WriteStreamFailureAsync(
                 response,
-                "response.created",
-                new
-                {
-                    type = "response.created",
-                    response = createdResponse,
-                    sequence_number = ++sequenceNumber,
-                },
-                ct);
+                normalized,
+                createdAt,
+                ++sequenceNumber,
+                completion.Error.Code,
+                completion.Error.Message,
+                CancellationToken.None);
+            return;
+        }
 
-            var outputItem = BuildOutputMessage(normalized.MessageItemId, "in_progress", text: null);
+        var completedText = completion.OutputText;
+        await WriteSseFrameAsync(
+            response,
+            "response.output_text.done",
+            new
+            {
+                type = "response.output_text.done",
+                item_id = normalized.MessageItemId,
+                output_index = 0,
+                content_index = 0,
+                text = completedText,
+                sequence_number = ++sequenceNumber,
+            },
+            ct);
+
+        var completedOutputItem = BuildOutputMessage(normalized.MessageItemId, "completed", completedText);
+        await WriteSseFrameAsync(
+            response,
+            "response.output_item.done",
+            new
+            {
+                type = "response.output_item.done",
+                output_index = 0,
+                item = completedOutputItem,
+                sequence_number = ++sequenceNumber,
+            },
+            ct);
+
+        var nextOutputIndex = 1;
+        foreach (var toolCall in completion.ForwardedToolCalls)
+        {
+            var functionCallItem = BuildFunctionCallOutputItem(toolCall);
             await WriteSseFrameAsync(
                 response,
                 "response.output_item.added",
                 new
                 {
                     type = "response.output_item.added",
-                    output_index = 0,
-                    item = outputItem,
+                    output_index = nextOutputIndex,
+                    item = functionCallItem,
                     sequence_number = ++sequenceNumber,
                 },
                 ct);
-
-            var completion = await completionService.StreamAsync(
-                provider,
-                request,
-                toolContextMetadata,
-                toolClassification,
-                async (delta, token) =>
-                {
-                    await WriteSseFrameAsync(
-                        response,
-                        "response.output_text.delta",
-                        new
-                        {
-                            type = "response.output_text.delta",
-                            item_id = normalized.MessageItemId,
-                            output_index = 0,
-                            content_index = 0,
-                            delta,
-                            sequence_number = ++sequenceNumber,
-                        },
-                        token);
-                },
-                ct);
-            usage = completion.Usage;
-
-            var completedText = completion.Text;
-            await WriteSseFrameAsync(
-                response,
-                "response.output_text.done",
-                new
-                {
-                    type = "response.output_text.done",
-                    item_id = normalized.MessageItemId,
-                    output_index = 0,
-                    content_index = 0,
-                    text = completedText,
-                    sequence_number = ++sequenceNumber,
-                },
-                ct);
-
-            var completedOutputItem = BuildOutputMessage(normalized.MessageItemId, "completed", completedText);
             await WriteSseFrameAsync(
                 response,
                 "response.output_item.done",
                 new
                 {
                     type = "response.output_item.done",
-                    output_index = 0,
-                    item = completedOutputItem,
+                    output_index = nextOutputIndex,
+                    item = functionCallItem,
                     sequence_number = ++sequenceNumber,
                 },
                 ct);
+            nextOutputIndex++;
+        }
 
-            var completedToolCalls = completion.ForwardedToolCalls;
-            await PersistForwardedToolCallsAsync(
-                responseSessionRegistrationPort,
-                logger,
-                responseSession,
-                toolClassification,
-                completedToolCalls,
-                DateTimeOffset.UtcNow,
-                ct);
-            await TryResolveIncomingToolResultsAsync(
-                responseSessionRegistrationPort,
-                logger,
-                previousSnapshot,
-                normalized,
-                ct);
+        var completedResponse = BuildCompletedResponse(
+            normalized,
+            createdAt,
+            DateTimeOffset.UtcNow.ToUnixTimeSeconds(),
+            completedText,
+            completion.ForwardedToolCalls,
+            completion.Usage is null ? null : MapUsage(completion.Usage));
 
-            var nextOutputIndex = 1;
-            foreach (var toolCall in completedToolCalls)
+        await WriteSseFrameAsync(
+            response,
+            "response.completed",
+            new
             {
-                var functionCallItem = BuildFunctionCallOutputItem(toolCall);
-                await WriteSseFrameAsync(
-                    response,
-                    "response.output_item.added",
-                    new
-                    {
-                        type = "response.output_item.added",
-                        output_index = nextOutputIndex,
-                        item = functionCallItem,
-                        sequence_number = ++sequenceNumber,
-                    },
-                    ct);
-                await WriteSseFrameAsync(
-                    response,
-                    "response.output_item.done",
-                    new
-                    {
-                        type = "response.output_item.done",
-                        output_index = nextOutputIndex,
-                        item = functionCallItem,
-                        sequence_number = ++sequenceNumber,
-                    },
-                    ct);
-                nextOutputIndex++;
-            }
-
-            var completedResponse = BuildCompletedResponse(
-                normalized,
-                createdAt,
-                DateTimeOffset.UtcNow.ToUnixTimeSeconds(),
-                completedText,
-                completedToolCalls,
-                usage is null ? null : MapUsage(usage));
-
-            await WriteSseFrameAsync(
-                response,
-                "response.completed",
-                new
-                {
-                    type = "response.completed",
-                    response = completedResponse,
-                    sequence_number = ++sequenceNumber,
-                },
-                ct);
-            await TryUpdateSessionStatusAsync(
-                responseSessionRegistrationPort,
-                logger,
-                responseSession,
-                LlmSessionStatus.Completed,
-                ct);
-        }
-        catch (NyxIdAuthenticationRequiredException ex)
-        {
-            await TryUpdateSessionStatusAsync(
-                responseSessionRegistrationPort,
-                logger,
-                responseSession,
-                LlmSessionStatus.Failed,
-                CancellationToken.None);
-            // NyxID authentication-required messages describe why the caller's
-            // token was rejected; surface verbatim (not server internals).
-            await WriteStreamFailureAsync(
-                response,
-                normalized,
-                createdAt,
-                ++sequenceNumber,
-                "authentication_required",
-                ex.Message,
-                ct);
-        }
-        catch (NyxIdUpstreamException ex)
-        {
-            await TryUpdateSessionStatusAsync(
-                responseSessionRegistrationPort,
-                logger,
-                responseSession,
-                LlmSessionStatus.Failed,
-                CancellationToken.None);
-            var correlation = LogAndCorrelate(logger, ex, "stream_nyxid_upstream", normalized.ResponseId);
-            await WriteStreamFailureAsync(
-                response,
-                normalized,
-                createdAt,
-                ++sequenceNumber,
-                ex.Kind.ToString().ToLowerInvariant(),
-                $"Upstream provider error. Correlation: {correlation}",
-                ct);
-        }
-        catch (OperationCanceledException) when (ct.IsCancellationRequested)
-        {
-            await TryUpdateSessionStatusAsync(
-                responseSessionRegistrationPort,
-                logger,
-                responseSession,
-                LlmSessionStatus.Cancelled,
-                CancellationToken.None);
-        }
-        catch (Exception ex)
-        {
-            await TryUpdateSessionStatusAsync(
-                responseSessionRegistrationPort,
-                logger,
-                responseSession,
-                LlmSessionStatus.Failed,
-                CancellationToken.None);
-            var correlation = LogAndCorrelate(logger, ex, "stream_execution", normalized.ResponseId);
-            await WriteStreamFailureAsync(
-                response,
-                normalized,
-                createdAt,
-                ++sequenceNumber,
-                "execution_failed",
-                $"Execution failed. Correlation: {correlation}",
-                ct);
-        }
+                type = "response.completed",
+                response = completedResponse,
+                sequence_number = ++sequenceNumber,
+            },
+            ct);
     }
 
     /// <summary>
@@ -864,58 +836,6 @@ internal static partial class ResponsesApiEndpoints
         };
     }
 
-    private static List<ChatMessage> BuildLlmMessages(
-        NormalizedResponsesRequest normalized,
-        LlmSessionSnapshot? previousSnapshot)
-    {
-        var messages = new List<ChatMessage>();
-        if (normalized.ToolResults.Count > 0 && previousSnapshot != null)
-        {
-            var toolCalls = BuildPreviousToolCalls(normalized, previousSnapshot);
-            if (toolCalls.Count > 0)
-            {
-                messages.Add(new ChatMessage
-                {
-                    Role = "assistant",
-                    ToolCalls = toolCalls,
-                });
-            }
-
-            foreach (var result in normalized.ToolResults)
-                messages.Add(ChatMessage.Tool(result.CallId, result.Output));
-        }
-
-        if (!string.IsNullOrWhiteSpace(normalized.Prompt))
-            messages.Add(ChatMessage.User(normalized.Prompt));
-
-        return messages;
-    }
-
-    private static IReadOnlyList<ToolCall> BuildPreviousToolCalls(
-        NormalizedResponsesRequest normalized,
-        LlmSessionSnapshot previousSnapshot)
-    {
-        var forwardedCalls = previousSnapshot.ForwardedToolCalls ?? [];
-        var callsById = forwardedCalls
-            .GroupBy(static call => call.CallId, StringComparer.Ordinal)
-            .ToDictionary(static group => group.Key, static group => group.First(), StringComparer.Ordinal);
-        var result = new List<ToolCall>();
-        foreach (var input in normalized.ToolResults)
-        {
-            if (!callsById.TryGetValue(input.CallId, out var call))
-                continue;
-
-            result.Add(new ToolCall
-            {
-                Id = call.CallId,
-                Name = call.ToolName,
-                ArgumentsJson = string.IsNullOrWhiteSpace(call.ArgumentsJson) ? "{}" : call.ArgumentsJson,
-            });
-        }
-
-        return result;
-    }
-
     private static ResponsesUsage MapUsage(TokenUsage usage) =>
         new()
         {
@@ -925,14 +845,6 @@ internal static partial class ResponsesApiEndpoints
             TotalTokens = usage.TotalTokens,
             OutputTokensDetails = new ResponsesOutputTokensDetails(),
         };
-
-    private static ResponsesApplicationToolDeclaration ToApplicationToolDeclaration(
-        ResponsesToolDeclaration declaration) =>
-        new(
-            declaration.Name,
-            declaration.Description,
-            declaration.ParametersJson,
-            declaration.SchemaHash);
 
     private static ResponsesOutputMessage BuildOutputMessage(string id, string status, string? text)
     {
@@ -980,206 +892,18 @@ internal static partial class ResponsesApiEndpoints
         return builder.ToString();
     }
 
-    private static async Task<IResult?> PersistIncomingToolResultsAsync(
-        ILlmSessionRegistrationPort responseSessionRegistrationPort,
-        LlmSessionSnapshot previousSnapshot,
-        NormalizedResponsesRequest normalized,
+    private static async Task WriteSseFrameAsync(
+        HttpResponse response,
+        string eventName,
+        object payload,
         CancellationToken ct)
     {
-        var callsById = (previousSnapshot.ForwardedToolCalls ?? [])
-            .GroupBy(static call => call.CallId, StringComparer.Ordinal)
-            .ToDictionary(static group => group.Key, static group => group.First(), StringComparer.Ordinal);
-
-        foreach (var result in normalized.ToolResults)
-        {
-            if (!callsById.TryGetValue(result.CallId, out var call))
-            {
-                return ToErrorResult(
-                    StatusCodes.Status400BadRequest,
-                    "tool_call_not_found",
-                    $"previous_response_id has no forwarded tool call '{result.CallId}'.");
-            }
-
-            var schemaHash = result.SchemaHash ?? call.SchemaHash;
-            if (!string.Equals(call.SchemaHash, schemaHash, StringComparison.Ordinal))
-            {
-                return ToErrorResult(
-                    StatusCodes.Status400BadRequest,
-                    "tool_schema_hash_mismatch",
-                    $"Forwarded tool call '{result.CallId}' schema hash mismatch.");
-            }
-
-            if (call.Status == LlmSessionForwardedToolCallStatus.Resolved)
-                continue;
-
-            if (call.Status is LlmSessionForwardedToolCallStatus.Cancelled
-                or LlmSessionForwardedToolCallStatus.Expired)
-            {
-                return ToErrorResult(
-                    StatusCodes.Status400BadRequest,
-                    "tool_call_not_available",
-                    $"Forwarded tool call '{result.CallId}' is {call.Status} and cannot receive a result.");
-            }
-
-            try
-            {
-                await responseSessionRegistrationPort.ReceiveForwardedToolResultAsync(
-                    previousSnapshot.ActorId,
-                    previousSnapshot.ResponseId,
-                    result.CallId,
-                    schemaHash,
-                    result.Output,
-                    ct);
-            }
-            catch (InvalidOperationException ex)
-            {
-                return ToErrorResult(
-                    StatusCodes.Status400BadRequest,
-                    "tool_result_rejected",
-                    ex.Message);
-            }
-        }
-
-        return null;
-    }
-
-    private static bool TryBuildAlreadyResolvedToolResultResponse(
-        NormalizedResponsesRequest normalized,
-        LlmSessionSnapshot previousSnapshot,
-        [NotNullWhen(true)] out IResult? result)
-    {
-        result = null;
-        if (normalized.ToolResults.Count == 0)
-            return false;
-
-        var callsById = (previousSnapshot.ForwardedToolCalls ?? [])
-            .GroupBy(static call => call.CallId, StringComparer.Ordinal)
-            .ToDictionary(static group => group.Key, static group => group.First(), StringComparer.Ordinal);
-        var resolvedOutputs = new List<string>();
-        foreach (var input in normalized.ToolResults)
-        {
-            if (!callsById.TryGetValue(input.CallId, out var call) ||
-                call.Status != LlmSessionForwardedToolCallStatus.Resolved)
-            {
-                return false;
-            }
-
-            var schemaHash = input.SchemaHash ?? call.SchemaHash;
-            if (!string.Equals(call.SchemaHash, schemaHash, StringComparison.Ordinal))
-            {
-                result = ToErrorResult(
-                    StatusCodes.Status400BadRequest,
-                    "tool_schema_hash_mismatch",
-                    $"Forwarded tool call '{input.CallId}' schema hash mismatch.");
-                return true;
-            }
-
-            resolvedOutputs.Add(string.IsNullOrWhiteSpace(call.ResultJson) ? input.Output : call.ResultJson!);
-        }
-
-        var now = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
-        var outputText = resolvedOutputs.Count == 1
-            ? resolvedOutputs[0]
-            : JsonSerializer.Serialize(resolvedOutputs, JsonOptions);
-        result = Results.Json(
-            BuildCompletedResponse(
-                normalized,
-                now,
-                now,
-                outputText,
-                [],
-                null),
-            JsonOptions,
-            statusCode: StatusCodes.Status200OK);
-        return true;
-    }
-
-    private static async Task TryResolveIncomingToolResultsAsync(
-        ILlmSessionRegistrationPort responseSessionRegistrationPort,
-        ILogger logger,
-        LlmSessionSnapshot? previousSnapshot,
-        NormalizedResponsesRequest normalized,
-        CancellationToken ct)
-    {
-        if (previousSnapshot is null || normalized.ToolResults.Count == 0)
-            return;
-
-        foreach (var callId in normalized.ToolResults
-                     .Select(static result => result.CallId)
-                     .Where(static callId => !string.IsNullOrWhiteSpace(callId))
-                     .Distinct(StringComparer.Ordinal))
-        {
-            try
-            {
-                await responseSessionRegistrationPort.ResolveForwardedToolResultAsync(
-                    previousSnapshot.ActorId,
-                    previousSnapshot.ResponseId,
-                    callId,
-                    ct);
-            }
-            catch (OperationCanceledException) when (ct.IsCancellationRequested)
-            {
-                throw;
-            }
-            catch (Exception ex)
-            {
-                logger.LogWarning(
-                    ex,
-                    "Failed to mark forwarded Responses tool call {CallId} as resolved for response {ResponseId}.",
-                    callId,
-                    previousSnapshot.ResponseId);
-            }
-        }
-    }
-
-    private static async Task PersistForwardedToolCallsAsync(
-        ILlmSessionRegistrationPort responseSessionRegistrationPort,
-        ILogger logger,
-        LlmSessionRegistrationResult responseSession,
-        ResponsesToolClassification toolClassification,
-        IReadOnlyList<ToolCall> toolCalls,
-        DateTimeOffset emittedAt,
-        CancellationToken ct)
-    {
-        if (toolCalls.Count == 0)
-            return;
-
-        var declarations = toolClassification.ForwardedTools.ToDictionary(static tool => tool.Name, StringComparer.Ordinal);
-        var expiry = emittedAt.AddHours(24);
-        foreach (var toolCall in toolCalls)
-        {
-            if (string.IsNullOrWhiteSpace(toolCall.Id))
-                throw new InvalidOperationException("Forwarded tool call is missing call_id.");
-            if (string.IsNullOrWhiteSpace(toolCall.Name))
-                throw new InvalidOperationException($"Forwarded tool call '{toolCall.Id}' is missing tool name.");
-            if (!declarations.TryGetValue(toolCall.Name, out var declaration))
-            {
-                throw new InvalidOperationException(
-                    $"Forwarded tool call '{toolCall.Id}' references undeclared tool '{toolCall.Name}'.");
-            }
-
-            var argumentsJson = string.IsNullOrWhiteSpace(toolCall.ArgumentsJson) ? "{}" : toolCall.ArgumentsJson;
-            var call = new LlmSessionForwardedToolCall
-            {
-                CallId = toolCall.Id,
-                ToolName = toolCall.Name,
-                SchemaHash = declaration.SchemaHash,
-                Arguments = ResponsesJsonValues.ParseBoundaryPayload(argumentsJson),
-                Status = LlmSessionForwardedToolCallStatus.Pending,
-                EmittedAt = Timestamp.FromDateTimeOffset(emittedAt),
-                Expiry = Timestamp.FromDateTimeOffset(expiry),
-            };
-
-            await responseSessionRegistrationPort.RecordForwardedToolCallAsync(
-                responseSession.ActorId,
-                responseSession.ResponseId,
-                call,
-                ct);
-            logger.LogDebug(
-                "Persisted forwarded Responses tool call {CallId} for response {ResponseId}.",
-                toolCall.Id,
-                responseSession.ResponseId);
-        }
+        var json = JsonSerializer.Serialize(payload, JsonOptions);
+        var bytes = Encoding.UTF8.GetBytes($"event: {eventName}\n");
+        await response.Body.WriteAsync(bytes, ct);
+        bytes = Encoding.UTF8.GetBytes($"data: {json}\n\n");
+        await response.Body.WriteAsync(bytes, ct);
+        await response.Body.FlushAsync(ct);
     }
 
     private static async Task WriteStreamFailureAsync(
@@ -1251,210 +975,6 @@ internal static partial class ResponsesApiEndpoints
             Usage = null,
             Metadata = new Dictionary<string, string>(StringComparer.Ordinal),
         };
-    }
-
-    internal static ResponsesToolProviderContext BuildToolProviderContext(
-        ResponsesCallerScope callerScope,
-        string responseId,
-        string bearerToken)
-    {
-        ArgumentNullException.ThrowIfNull(callerScope);
-
-        return new ResponsesToolProviderContext(
-            new ResponsesToolProviderCallerScope(
-                callerScope.ScopeId,
-                callerScope.OwnerSubject,
-                callerScope.OriginKind.ToString()),
-            new Dictionary<string, string>(StringComparer.Ordinal)
-            {
-                [LLMRequestMetadataKeys.RequestId] = responseId,
-                [LLMRequestMetadataKeys.ResponseId] = responseId,
-                [LLMRequestMetadataKeys.ScopeId] = callerScope.ScopeId,
-                [LLMRequestMetadataKeys.OwnerSubject] = callerScope.OwnerSubject,
-                [ChannelMetadataKeys.RegistrationScopeId] = callerScope.ScopeId,
-                [LLMRequestMetadataKeys.NyxIdAccessToken] = bearerToken,
-            });
-    }
-
-    internal static async Task<ChatRouteDecision> ResolveResponsesChatRouteAsync(
-        IChatRoutePolicyQueryPort queryPort,
-        ChatRouteResolver resolver,
-        ResponsesCallerScope callerScope,
-        string model,
-        ToolMode toolMode,
-        string contentHint,
-        CancellationToken ct)
-    {
-        var ownerScope = OwnerScope.ForNyxIdNative(callerScope.ScopeId);
-        var snapshot = await queryPort.LookupForCallerAsync(ownerScope, ct);
-        return resolver.Resolve(snapshot, new ChatRouteInput
-        {
-            SourceKind = ChatSourceKind.NyxResponses,
-            CallerScope = new ChatRouteCallerScope
-            {
-                NyxUserId = ownerScope.NyxUserId,
-                Platform = ownerScope.Platform,
-                RegistrationScopeId = ownerScope.RegistrationScopeId,
-                SenderId = ownerScope.SenderId,
-            },
-            Channel = string.Empty,
-            CommandName = string.Empty,
-            ContentHint = contentHint,
-            ToolMode = toolMode,
-            Model = model,
-        });
-    }
-
-    internal static ToolMode ResolveToolMode(int declaredToolCount, int inlineToolResultCount)
-    {
-        if (inlineToolResultCount > 0)
-            return ToolMode.Inline;
-        return declaredToolCount > 0 ? ToolMode.Declared : ToolMode.None;
-    }
-
-    internal static string BuildContentHint(string? content)
-    {
-        var normalized = content?.Trim();
-        if (string.IsNullOrWhiteSpace(normalized))
-            return string.Empty;
-        const int maxContentHintLength = 160;
-        return normalized.Length <= maxContentHintLength
-            ? normalized
-            : normalized[..maxContentHintLength];
-    }
-
-    private static LlmSessionRecord BuildResponseSessionRecord(
-        NormalizedResponsesRequest normalized,
-        ResponsesCallerScope callerScope,
-        DateTimeOffset createdAt)
-    {
-        return new LlmSessionRecord
-        {
-            ResponseId = normalized.ResponseId,
-            ScopeId = callerScope.ScopeId,
-            OwnerSubject = callerScope.OwnerSubject,
-            OriginKind = callerScope.OriginKind,
-            PreviousResponseId = normalized.PreviousResponseId ?? string.Empty,
-            Status = LlmSessionStatus.Accepted,
-            CreatedAt = Timestamp.FromDateTime(createdAt.UtcDateTime),
-            UpdatedAt = Timestamp.FromDateTime(createdAt.UtcDateTime),
-            Ttl = Duration.FromTimeSpan(TimeSpan.FromHours(24)),
-        };
-    }
-
-    private static IResult? ValidatePreviousResponse(
-        LlmSessionSnapshot? previous,
-        ResponsesCallerScope callerScope)
-    {
-        var visibilityError = ValidateResponseVisibility(
-            previous,
-            callerScope,
-            "previous_response_not_found",
-            "previous_response_id does not refer to a visible response session.");
-        if (visibilityError is not null)
-            return visibilityError;
-
-        var visiblePrevious = previous!;
-        if (visiblePrevious.Ttl > TimeSpan.Zero &&
-            visiblePrevious.CreatedAt.Add(visiblePrevious.Ttl) <= DateTimeOffset.UtcNow)
-        {
-            return ToErrorResult(
-                StatusCodes.Status400BadRequest,
-                "previous_response_expired",
-                "previous_response_id refers to an expired response session.");
-        }
-
-        if (visiblePrevious.Status is LlmSessionStatus.Cancelled
-            or LlmSessionStatus.Expired
-            or LlmSessionStatus.Failed)
-        {
-            return ToErrorResult(
-                StatusCodes.Status400BadRequest,
-                "previous_response_not_available",
-                "previous_response_id refers to a response session that cannot be continued.");
-        }
-
-        return null;
-    }
-
-    private static IResult? ValidateResponseVisibility(
-        LlmSessionSnapshot? response,
-        ResponsesCallerScope callerScope,
-        string notFoundCode,
-        string notFoundMessage)
-    {
-        if (response is null)
-        {
-            return ToErrorResult(
-                StatusCodes.Status404NotFound,
-                notFoundCode,
-                notFoundMessage);
-        }
-
-        if (!string.Equals(response.ScopeId, callerScope.ScopeId, StringComparison.Ordinal) ||
-            !string.Equals(response.OwnerSubject, callerScope.OwnerSubject, StringComparison.Ordinal))
-        {
-            return ToErrorResult(
-                StatusCodes.Status403Forbidden,
-                "response_scope_mismatch",
-                "response id is not visible to the current caller scope.");
-        }
-
-        if (response.OriginKind != callerScope.OriginKind)
-        {
-            return ToErrorResult(
-                StatusCodes.Status403Forbidden,
-                "response_origin_mismatch",
-                "response id origin does not match the current ingress origin.");
-        }
-
-        return null;
-    }
-
-    private static async Task TryUpdateSessionStatusAsync(
-        ILlmSessionRegistrationPort responseSessionRegistrationPort,
-        ILogger logger,
-        LlmSessionRegistrationResult responseSession,
-        LlmSessionStatus status,
-        CancellationToken ct)
-    {
-        try
-        {
-            await responseSessionRegistrationPort.UpdateStatusAsync(
-                responseSession.ActorId,
-                responseSession.ResponseId,
-                status,
-                ct);
-        }
-        catch (OperationCanceledException) when (ct.IsCancellationRequested)
-        {
-            throw;
-        }
-        catch (Exception ex)
-        {
-            // The response session has already been accepted. Completion markers are
-            // observable state, but they must not leak persistence failures or secrets
-            // into the Responses payload path.
-            logger.LogWarning(
-                ex,
-                "Failed to update response session {ResponseId} to {Status}.",
-                responseSession.ResponseId,
-                status);
-        }
-    }
-
-    private static async Task WriteSseFrameAsync(
-        HttpResponse response,
-        string eventName,
-        object payload,
-        CancellationToken ct)
-    {
-        var json = JsonSerializer.Serialize(payload, JsonOptions);
-        var bytes = Encoding.UTF8.GetBytes($"event: {eventName}\n");
-        await response.Body.WriteAsync(bytes, ct);
-        bytes = Encoding.UTF8.GetBytes($"data: {json}\n\n");
-        await response.Body.WriteAsync(bytes, ct);
-        await response.Body.FlushAsync(ct);
     }
 
     private static IResult ToErrorResult(int statusCode, string code, string message) =>
