+++ src/platform/Aevatar.GAgentService.Application/Responses/ResponsesIds.cs
@@ -1,0 +1,17 @@
+using System.Security.Cryptography;
+
+namespace Aevatar.GAgentService.Application.Responses;
+
+public static class ResponsesIds
+{
+    public static string NewResponseId() => "resp_" + NewOpaqueId();
+
+    public static string NewMessageId() => "msg_" + NewOpaqueId();
+
+    public static string NewOpaqueId()
+    {
+        Span<byte> bytes = stackalloc byte[12];
+        RandomNumberGenerator.Fill(bytes);
+        return Convert.ToHexString(bytes).ToLowerInvariant();
+    }
+}
